Add command to remove all past dates and events
It will search for all dates within the past and remove them. Afterwards
all events with no dates are searched and removed.

As DataHandler is used for deletion of events, there is already logging
within TYPO3, as well as recursive deletion for sys_file_references.

To speedup for large data sets, deletion of dates is done without
DataHandler.