Skip import of events with file inconsistency.
Those would previously lead to a broken import, as TYPO3 would throw
an unhandled
exception.

This is now catched, and converted to a warning log entry instead.

Do not break import for file inconsistency

TYPO3 will verify the consistency of added files.
Files of import might be of a different mime type than their file
extension.
This will result in an exception.

Relates: #80, #12182